% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvcrand_package.R
\docType{package}
\name{cvcrand}
\alias{cvcrand}
\title{cvcrand: a package for efficient design and analysis of cluster randomized trials}
\description{
cvcrand: A Package for Covariate-constrained Randomization and the Clustered Permutation Test for
Cluster Randomized Trials
}
\section{cvcrand functions}{

The cvrall function performs constrained randomization on cluster randomized trials (CRTs)
The cvrcov function performs covariate-constrained randomization on cluster randomized trials (CRTs)
The cptest function performs permutation test on the outcome of cluster randomized trials (CRTs)
}

